extract licence and privacy into their own pages
