+++ src/pages/Privacy.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import Helmet from 'react-helmet';
 import { Link } from 'react-router-dom';
-import './Terms.scss';
 
 const Privacy = () => (
   <div className="terms">
