+++ src/pages/Terms/index.js
